CHE-5725: code refactoring
