+++ plugins/plugin-testing-java/plugin-testing-junit/che-plugin-testing-junit-ide/src/main/java/org/eclipse/che/plugin/testing/junit/ide/action/DebugJUnitTestAction.java
@@ -13,44 +13,29 @@ package org.eclipse.che.plugin.testing.junit.ide.action;
 import com.google.inject.Inject;
 import com.google.web.bindery.event.shared.EventBus;
 
-import org.eclipse.che.api.core.jsonrpc.commons.JsonRpcError;
-import org.eclipse.che.api.core.jsonrpc.commons.JsonRpcPromise;
-import org.eclipse.che.api.testing.shared.TestExecutionContext;
-import org.eclipse.che.api.testing.shared.TestLaunchResult;
 import org.eclipse.che.ide.api.action.ActionEvent;
+import org.eclipse.che.ide.api.action.Presentation;
 import org.eclipse.che.ide.api.app.AppContext;
-import org.eclipse.che.ide.api.debug.DebugConfiguration;
 import org.eclipse.che.ide.api.debug.DebugConfigurationsManager;
 import org.eclipse.che.ide.api.notification.NotificationManager;
-import org.eclipse.che.ide.api.notification.StatusNotification;
 import org.eclipse.che.ide.dto.DtoFactory;
 import org.eclipse.che.ide.util.Pair;
 import org.eclipse.che.plugin.testing.ide.TestServiceClient;
+import org.eclipse.che.plugin.testing.ide.action.RunDebugTestAbstractAction;
 import org.eclipse.che.plugin.testing.ide.handler.TestingHandler;
-import org.eclipse.che.plugin.testing.ide.model.GeneralTestingEventsProcessor;
 import org.eclipse.che.plugin.testing.ide.view2.TestResultPresenter;
 import org.eclipse.che.plugin.testing.junit.ide.JUnitTestLocalizationConstant;
 import org.eclipse.che.plugin.testing.junit.ide.JUnitTestResources;
 
-import static java.lang.Boolean.TRUE;
-import static java.util.Collections.emptyMap;
+import javax.validation.constraints.NotNull;
+
 import static java.util.Collections.singletonList;
-import static org.eclipse.che.ide.api.notification.StatusNotification.DisplayMode.FLOAT_MODE;
-import static org.eclipse.che.ide.api.notification.StatusNotification.Status.FAIL;
-import static org.eclipse.che.ide.api.notification.StatusNotification.Status.PROGRESS;
-import static org.eclipse.che.ide.api.notification.StatusNotification.Status.SUCCESS;
 import static org.eclipse.che.ide.workspace.perspectives.project.ProjectPerspective.PROJECT_PERSPECTIVE_ID;
 
 /**
  * The action for activation debugger for JUnit test.
  */
-public class DebugJUnitTestAction extends JUnitAbstractAction {
-    private       TestServiceClient          client;
-    private       TestingHandler             testingHandler;
-    private final NotificationManager        notificationManager;
-    private       DebugConfigurationsManager debugConfigurationsManager;
-    private final TestResultPresenter        testResultPresenter;
-
+public class DebugJUnitTestAction extends RunDebugTestAbstractAction {
     @Inject
     public DebugJUnitTestAction(JUnitTestResources resources,
                                 EventBus eventBus,
@@ -63,6 +48,9 @@ public class DebugJUnitTestAction extends JUnitAbstractAction {
                                 TestResultPresenter testResultPresenter,
                                 JUnitTestLocalizationConstant localization) {
         super(eventBus,
+              testResultPresenter,
+              testingHandler,
+              debugConfigurationsManager,
               client,
               dtoFactory,
               appContext,
@@ -71,60 +59,22 @@ public class DebugJUnitTestAction extends JUnitAbstractAction {
               localization.actionDebugDescription(),
               localization.actionDebugTestTitle(),
               resources.testIcon());
-        this.client = client;
-        this.testingHandler = testingHandler;
-        this.notificationManager = notificationManager;
-        this.debugConfigurationsManager = debugConfigurationsManager;
-        this.testResultPresenter = testResultPresenter;
     }
 
     @Override
     public void actionPerformed(ActionEvent e) {
-        final StatusNotification notification = new StatusNotification("Debugging Tests...", PROGRESS, FLOAT_MODE);
-        notificationManager.notify(notification);
-
-        Pair<String, String> frameworkAndTestName = Pair.of(JUNIT_TEST_FRAMEWORK, null);
-        TestExecutionContext context = createTestExecutionContext(frameworkAndTestName, contextType, selectedNodePath);
-        context.withDebugModeEnable(TRUE);
-
-        GeneralTestingEventsProcessor eventsProcessor = new GeneralTestingEventsProcessor(JUNIT_TEST_FRAMEWORK,
-                                                                                          testResultPresenter.getRootState());
-        testingHandler.setProcessor(eventsProcessor);
-        eventsProcessor.addListener(testResultPresenter.getEventListener());
-
-        JsonRpcPromise<TestLaunchResult> testResultPromise = client.runTests(context);
-        testResultPromise.onSuccess(result -> onTestRanSuccessfully(result, eventsProcessor, notification))
-                         .onFailure(exception -> onTestRanFailed(exception, notification));
+        Pair<String, String> frameworkAndTestName = Pair.of(JUNIT_FRAMEWORK_NAME, null);
+        actionPerformed(frameworkAndTestName, true);
     }
 
-    private void onTestRanSuccessfully(TestLaunchResult result,
-                                       GeneralTestingEventsProcessor eventsProcessor,
-                                       StatusNotification notification) {
-        notification.setStatus(SUCCESS);
-        if (result.isSuccess()) {
-            notification.setTitle("Test runner executed successfully.");
-            testResultPresenter.handleResponse();
-            runRemoteJavaDebugger(eventsProcessor, result.getDebugPort());
-        } else {
-            notification.setTitle("Test runner failed to execute.");
+    @Override
+    public void updateInPerspective(@NotNull ActionEvent e) {
+        Presentation presentation = e.getPresentation();
+        presentation.setVisible(!isEditorInFocus);
+        if (!isEditorInFocus) {
+            analyzeProjectTreeSelection();
         }
+        presentation.setEnabled(isEnable);
     }
 
-    private void runRemoteJavaDebugger(GeneralTestingEventsProcessor eventsProcessor, int port) {
-        DebugConfiguration debugger = debugConfigurationsManager.createConfiguration("jdb",
-                                                                                     JUNIT_TEST_FRAMEWORK,
-                                                                                     "localhost",
-                                                                                     port,
-                                                                                     emptyMap());
-        eventsProcessor.setDebuggerConfiguration(debugger, debugConfigurationsManager);
-
-        debugConfigurationsManager.apply(debugger);
-    }
-
-    private void onTestRanFailed(JsonRpcError exception, StatusNotification notification) {
-        final String errorMessage = (exception.getMessage() != null) ? exception.getMessage()
-                                                                     : "Failed to run test cases";
-        notification.setContent(errorMessage);
-        notification.setStatus(FAIL);
-    }
 }
