+++ plugins/plugin-testing-java/plugin-testing-testng/che-plugin-testing-testng-ide/src/main/java/org/eclipse/che/plugin/testing/testng/ide/action/RunTestNgTestAction.java
@@ -13,43 +13,33 @@ package org.eclipse.che.plugin.testing.testng.ide.action;
 import com.google.inject.Inject;
 import com.google.web.bindery.event.shared.EventBus;
 
-import org.eclipse.che.api.core.jsonrpc.commons.JsonRpcError;
-import org.eclipse.che.api.core.jsonrpc.commons.JsonRpcPromise;
-import org.eclipse.che.api.testing.shared.TestExecutionContext;
-import org.eclipse.che.api.testing.shared.TestLaunchResult;
 import org.eclipse.che.ide.api.action.ActionEvent;
+import org.eclipse.che.ide.api.action.Presentation;
 import org.eclipse.che.ide.api.app.AppContext;
+import org.eclipse.che.ide.api.debug.DebugConfigurationsManager;
 import org.eclipse.che.ide.api.notification.NotificationManager;
-import org.eclipse.che.ide.api.notification.StatusNotification;
 import org.eclipse.che.ide.dto.DtoFactory;
 import org.eclipse.che.ide.util.Pair;
 import org.eclipse.che.plugin.testing.ide.TestServiceClient;
+import org.eclipse.che.plugin.testing.ide.action.RunDebugTestAbstractAction;
 import org.eclipse.che.plugin.testing.ide.handler.TestingHandler;
-import org.eclipse.che.plugin.testing.ide.model.GeneralTestingEventsProcessor;
 import org.eclipse.che.plugin.testing.ide.view2.TestResultPresenter;
 import org.eclipse.che.plugin.testing.testng.ide.TestNgLocalizationConstant;
 import org.eclipse.che.plugin.testing.testng.ide.TestNgResources;
 
-import static java.lang.Boolean.FALSE;
+import javax.validation.constraints.NotNull;
+
 import static java.util.Collections.singletonList;
-import static org.eclipse.che.ide.api.notification.StatusNotification.DisplayMode.FLOAT_MODE;
-import static org.eclipse.che.ide.api.notification.StatusNotification.Status.FAIL;
-import static org.eclipse.che.ide.api.notification.StatusNotification.Status.PROGRESS;
-import static org.eclipse.che.ide.api.notification.StatusNotification.Status.SUCCESS;
 import static org.eclipse.che.ide.workspace.perspectives.project.ProjectPerspective.PROJECT_PERSPECTIVE_ID;
 
 /**
  * Action for running TestNg test.
  */
-public class RunTestNgTestAction extends TestNGAbstractAction {
-    private final TestServiceClient   client;
-    private final TestingHandler      testingHandler;
-    private final NotificationManager notificationManager;
-    private final TestResultPresenter testResultPresenter;
-
+public class RunTestNgTestAction extends RunDebugTestAbstractAction {
     @Inject
     public RunTestNgTestAction(TestNgResources resources,
                                EventBus eventBus,
+                               DebugConfigurationsManager debugConfigurationsManager,
                                TestServiceClient client,
                                TestingHandler testingHandler,
                                DtoFactory dtoFactory,
@@ -58,6 +48,9 @@ public class RunTestNgTestAction extends TestNGAbstractAction {
                                TestResultPresenter testResultPresenter,
                                TestNgLocalizationConstant localization) {
         super(eventBus,
+              testResultPresenter,
+              testingHandler,
+              debugConfigurationsManager,
               client,
               dtoFactory,
               appContext,
@@ -66,45 +59,21 @@ public class RunTestNgTestAction extends TestNGAbstractAction {
               localization.actionRunTestDescription(),
               localization.actionRunTestTitle(),
               resources.testIcon());
-        this.client = client;
-        this.testingHandler = testingHandler;
-        this.notificationManager = notificationManager;
-        this.testResultPresenter = testResultPresenter;
     }
 
     @Override
     public void actionPerformed(ActionEvent e) {
-        final StatusNotification notification = new StatusNotification("Running Tests...", PROGRESS, FLOAT_MODE);
-        notificationManager.notify(notification);
-
-        Pair<String, String> frameworkAndTestName = Pair.of(TESTNG_TEST_FRAMEWORK, null);
-        TestExecutionContext context = createTestExecutionContext(frameworkAndTestName, contextType, selectedNodePath);
-        context.withDebugModeEnable(FALSE);
-
-        GeneralTestingEventsProcessor eventsProcessor = new GeneralTestingEventsProcessor(TESTNG_TEST_FRAMEWORK,
-                                                                                          testResultPresenter.getRootState());
-        testingHandler.setProcessor(eventsProcessor);
-        eventsProcessor.addListener(testResultPresenter.getEventListener());
-
-        JsonRpcPromise<TestLaunchResult> testResultPromise = client.runTests(context);
-        testResultPromise.onSuccess(result -> onTestRanSuccessfully(result, notification))
-                         .onFailure(exception -> onTestRanFailed(exception, notification));
+        Pair<String, String> frameworkAndTestName = Pair.of(TESTNG_FRAMEWORK_NAME, null);
+        actionPerformed(frameworkAndTestName, false);
     }
 
-    private void onTestRanSuccessfully(TestLaunchResult result, StatusNotification notification) {
-        notification.setStatus(SUCCESS);
-        if (result.isSuccess()) {
-            notification.setTitle("Test runner executed successfully.");
-            testResultPresenter.handleResponse();
-        } else {
-            notification.setTitle("Test runner failed to execute.");
+    @Override
+    public void updateInPerspective(@NotNull ActionEvent e) {
+        Presentation presentation = e.getPresentation();
+        presentation.setVisible(!isEditorInFocus);
+        if (!isEditorInFocus) {
+            analyzeProjectTreeSelection();
         }
-    }
-
-    private void onTestRanFailed(JsonRpcError exception, StatusNotification notification) {
-        final String errorMessage = (exception.getMessage() != null) ? exception.getMessage()
-                                                                     : "Failed to run test cases";
-        notification.setContent(errorMessage);
-        notification.setStatus(FAIL);
+        presentation.setEnabled(isEnable);
     }
 }
