+++ plugins/plugin-testing/che-plugin-testing-ide/pom.xml
@@ -70,10 +70,6 @@
         </dependency>
         <dependency>
             <groupId>org.eclipse.che.core</groupId>
-            <artifactId>che-core-api-project-shared</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.eclipse.che.core</groupId>
             <artifactId>che-core-api-testing-shared</artifactId>
         </dependency>
         <dependency>
